fix: move test utils to testing0utils dir
